pq_system/requirements.txt
# 基础框架
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==2.3.7
SQLAlchemy==2.0.23
python-dotenv==1.0.0

# 文件处理
pypdf2==3.0.1
python-pptx==0.6.21
markdown==3.4.4
html2text==2020.1.16
striprtf==0.0.22
pytesseract==0.3.10
pdf2image==1.16.3
moviepy==1.0.3
SpeechRecognition==3.10.0
pydub==0.25.1

# AI API
openai==1.3.5
requests==2.31.0

# 安全和工具
bcrypt==4.0.1
email-validator==2.0.0
Pillow==10.0.1
tqdm==4.66.1

# 前端
Bootstrap-Flask==2.3.0 